Rename BaseObservableList to TypeScript
