+++ src/observable/list/BaseMappedList.js
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {BaseObservableList} from "./BaseObservableList.js";
+import {BaseObservableList} from "./BaseObservableList";
 import {findAndUpdateInArray} from "./common.js";
 
 export class BaseMappedList extends BaseObservableList {
