+++ src/observable/list/ConcatList.js
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {BaseObservableList} from "./BaseObservableList.js";
+import {BaseObservableList} from "./BaseObservableList";
 
 export class ConcatList extends BaseObservableList {
     constructor(...sourceLists) {
